disable mixin.perf.reduce_blockstate_cache_rebuilds in ModernFix
Is incompatible with Rainbow Wood + Terrestria/the-other-biome-mod-by-the-same-guys for some reason.
Was probably what caused the crash with MoreCulling because it said something about like a wall block not having some property it should or smt and the incompatibility here is basically the server sending (or client receiving?) completely random blocks and states for the blocks in the modded biomes.